Release/v0.45.7 (#2393)
* SCRUM-5666 add GenomeLocation to the gene object

* correct json view

* correct josn view

* checkstyle
