+++ src/main/java/org/alliancegenome/curation_api/model/entities/associations/VariantGenomicLocationAssociation.java
@@ -19,11 +19,13 @@ import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexedEmb
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexingDependency;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.KeywordField;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonView;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.MappedSuperclass;
+import jakarta.persistence.Transient;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -45,50 +47,75 @@ public abstract class VariantGenomicLocationAssociation extends VariantLocationA
 	@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.SHALLOW)
 	@Fetch(FetchMode.JOIN)
 	private AssemblyComponent variantGenomicLocationAssociationObject;
-	
+
 	@FullTextField(analyzer = "autocompleteAnalyzer", searchAnalyzer = "autocompleteSearchAnalyzer")
 	@KeywordField(name = "variationStrand_keyword", aggregable = Aggregable.YES, sortable = Sortable.YES, searchable = Searchable.YES, normalizer = "sortNormalizer")
 	@JsonView({ CurationView.FieldsOnly.class })
 	@Column(length = 1)
 	private String variationStrand;
-	
+
 	@GenericField(projectable = Projectable.YES, sortable = Sortable.YES)
 	@JsonView({ CurationView.FieldsOnly.class })
 	private Integer numberAdditionalDnaBasePairs;
-	
+
 	@GenericField(projectable = Projectable.YES, sortable = Sortable.YES)
 	@JsonView({ CurationView.FieldsOnly.class })
 	private Integer numberRemovedDnaBasePairs;
-	
+
 	@FullTextField(analyzer = "autocompleteAnalyzer", searchAnalyzer = "autocompleteSearchAnalyzer")
 	@KeywordField(name = "paddedBase_keyword", aggregable = Aggregable.YES, sortable = Sortable.YES, searchable = Searchable.YES, normalizer = "sortNormalizer")
 	@JsonView({ CurationView.FieldsOnly.class })
 	@Column(length = 1)
 	private String paddedBase;
-	
+
 	@FullTextField(analyzer = "autocompleteAnalyzer", searchAnalyzer = "autocompleteSearchAnalyzer")
 	@KeywordField(name = "insertedSequence_keyword", aggregable = Aggregable.YES, sortable = Sortable.YES, searchable = Searchable.YES, normalizer = "sortNormalizer")
 	@JsonView({ CurationView.FieldsOnly.class })
 	@Column(columnDefinition = "TEXT")
 	private String insertedSequence;
-	
+
 	@FullTextField(analyzer = "autocompleteAnalyzer", searchAnalyzer = "autocompleteSearchAnalyzer")
 	@KeywordField(name = "deletedSequence_keyword", aggregable = Aggregable.YES, sortable = Sortable.YES, searchable = Searchable.YES, normalizer = "sortNormalizer")
 	@JsonView({ CurationView.FieldsOnly.class })
 	@Column(columnDefinition = "TEXT")
 	private String deletedSequence;
-	
+
 	@IndexedEmbedded(includePaths = {"curie", "name", "secondaryIdentifiers", "synonyms.name", "namespace",
 			"curie_keyword", "name_keyword", "secondaryIdentifiers_keyword", "synonyms.name_keyword", "namespace_keyword" })
 	@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.SHALLOW)
 	@ManyToOne
 	@JsonView({ CurationView.FieldsOnly.class })
 	private SOTerm dnaMutationType;
-	
+
 	@IndexedEmbedded(includePaths = {"curie", "name", "secondaryIdentifiers", "synonyms.name", "namespace",
 			"curie_keyword", "name_keyword", "secondaryIdentifiers_keyword", "synonyms.name_keyword", "namespace_keyword" })
 	@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.SHALLOW)
 	@ManyToOne
 	@JsonView({ CurationView.FieldsOnly.class })
 	private SOTerm geneLocalizationType;
+
+
+	@Transient
+	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
+	@JsonView({CurationView.FieldsOnly.class, CurationView.VariantDocument.class})
+	public String getNucleotideChange() {
+		if (getVariantAssociationSubject() != null && getVariantAssociationSubject().getVariantType() != null) {
+			String variantTypeCurie = getVariantAssociationSubject().getVariantType().getCurie();
+			if ("SO:0000667".equals(variantTypeCurie)) {
+				// Insertion
+				return "c>c" + getVariantSequence();
+			} else if ("SO:1000008".equals(variantTypeCurie)) {
+				// Point mutation
+				return getReferenceSequence() + ">" + getVariantSequence();
+			} else if ("SO:0000159".equals(variantTypeCurie)) {
+				// Deletion
+				return "t" + getReferenceSequence() + ">t";
+			}
+			if (getReferenceSequence() != null && getVariantSequence() != null) {
+				return getReferenceSequence() + ">" + getVariantSequence();
+			}
+		}
+		return null;
+	}
+
 }
